perf(text-encode): Bun::byteLength returns std::optional<double>, forcing a size_t→double→int32 round-trip when boxing the result of every Buffer.byteLength(chunk, 'utf8') call

+++ src/jsc/bindings/JSBuffer.h
@@ -37,7 +37,7 @@ extern "C" bool JSBuffer__isBuffer(JSC::JSGlobalObject*, JSC::EncodedJSValue);
 
 namespace Bun {
 
-std::optional<double> byteLength(JSC::JSString* str, WebCore::BufferEncodingType encoding);
+std::optional<size_t> byteLength(JSC::JSString* str, WebCore::BufferEncodingType encoding);
 
 namespace Buffer {
